Update dependencies and integrate SweetAlert2 for improved user notifications in division and team management

+++ src/pages/admin/division/DeleteDivision.jsx
@@ -1,15 +1,36 @@
 import React from "react";
 import ApiService from "../../../services/api-service";
 import DeleteIcon from "@mui/icons-material/Delete";
+import Swal from "sweetalert2";
 const DeleteDivision = ({ id }) => {
   const handleDelete = async () => {
-    try {
-      const response = await ApiService.delete(`divisions/${id}`);
-      console.log("Division deleted successfully:", response.data);
-      alert("Division deleted successfully");
-    } catch (error) {
-      console.error("Error deleting division:", error);
-    }
+   try {
+
+            Swal.fire({
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, delete it!'
+            }).then(async (result) => {
+                if (result.isConfirmed) {
+                    const response = await ApiService.delete(`divisions/${id}`);
+                    if(response.status === 200) {
+                        Swal.fire(
+                            'Deleted!',
+                            'Your file has been deleted.',
+                            'success'
+                        )
+                    }
+                }
+                
+            })
+           
+        } catch (error) {
+            console.error('Error deleting team:', error);
+        }
   };
   return (
     <div>
